ml/pyproject.toml
[project]
name = "track-em-toys-ml"
version = "0.1.0"
description = "ML training pipeline for toy image classification"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "coremltools>=7.2",
    "numpy>=1.24.0",
    "onnx>=1.16.0",
    "onnxruntime>=1.18.0",
    "scikit-learn>=1.4.0",
    "Pillow>=10.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
